Refactor component imports to align with new directory structure and improve code organization
